Module IR.ANF

type variable = int

val pp_variable : Stdlib.Format.formatter -> int -> unit

type label = int

val pp_label : Stdlib.Format.formatter -> int -> unit

type func_name = Syntax.ParseTree.func_name

val pp_func_name : Ppx_deriving_runtime.Format.formatter -> func_name -> Ppx_deriving_runtime.unit

val show_func_name : func_name -> Ppx_deriving_runtime.string

type adt_label = Syntax.ParseTree.adt_label

val pp_adt_label : Ppx_deriving_runtime.Format.formatter -> adt_label -> Ppx_deriving_runtime.unit

val show_adt_label : adt_label -> Ppx_deriving_runtime.string

type typ_name = Syntax.ParseTree.typ_name

val pp_typ_name : Ppx_deriving_runtime.Format.formatter -> typ_name -> Ppx_deriving_runtime.unit

val show_typ_name : typ_name -> Ppx_deriving_runtime.string

type intf_name = Syntax.ParseTree.intf_name

val pp_intf_name : Ppx_deriving_runtime.Format.formatter -> intf_name -> Ppx_deriving_runtime.unit

val show_intf_name : intf_name -> Ppx_deriving_runtime.string

type unary_op = Syntax.ParseTree.unary_op

val pp_unary_op : Ppx_deriving_runtime.Format.formatter -> unary_op -> Ppx_deriving_runtime.unit

val show_unary_op : unary_op -> Ppx_deriving_runtime.string

type binary_op = Syntax.ParseTree.binary_op

val pp_binary_op : Ppx_deriving_runtime.Format.formatter -> binary_op -> Ppx_deriving_runtime.unit

val show_binary_op : binary_op -> Ppx_deriving_runtime.string

type span = Syntax.SyntaxError.src_span

val pp_span : Ppx_deriving_runtime.Format.formatter -> span -> Ppx_deriving_runtime.unit

val show_span : span -> Ppx_deriving_runtime.string

type mutability =

| Imm

| Mut

val pp_mutability : Ppx_deriving_runtime.Format.formatter -> mutability -> Ppx_deriving_runtime.unit

val show_mutability : mutability -> Ppx_deriving_runtime.string

type permission =

| Owner

| Borrow of mutability

val pp_permission : Ppx_deriving_runtime.Format.formatter -> permission -> Ppx_deriving_runtime.unit

val show_permission : permission -> Ppx_deriving_runtime.string

type data_kind =

| Tuple of int

| Struct of typ_name

| ADT of typ_name * adt_label

| Impl of intf_name
(*
method dict/vtable of interface
*)

data_kind is used to classify different kinds of aggregate data. In the ANF IR, different kinds of aggregate data share a uniform representation.

val pp_data_kind : Ppx_deriving_runtime.Format.formatter -> data_kind -> Ppx_deriving_runtime.unit

val show_data_kind : data_kind -> Ppx_deriving_runtime.string

type lvalue = {

lv_var : variable;

lv_path : path;

lv_src : span;

}

lvalue is a value that may appear at the left hand side of an assignment. So every lvalue corresponds to a memory address. Some examples: point.x, *ptr

and path = path_node list

and path_node =

| Field of int

| FieldByName of string
(*
unresolved field access, resolved in later pass
*)

| Deref

| AsTag of int

| Tag

| Method of string

path_node is used to select a part of an existing lvalue:

Field k selects the k-th field from an aggregate, i.e. a tuple, a struct, an ADT or a method dictionary

Deref selects the address a pointer points to

AsTag l selects the data associated with label l in an ADT

Tag selects the tag/label of an ADT value

Method m selects the implementation of method m from an interface implementation

val show_lvalue : lvalue -> Ppx_deriving_runtime.string

val pp_path : Ppx_deriving_runtime.Format.formatter -> path -> Ppx_deriving_runtime.unit

val show_path : path -> Ppx_deriving_runtime.string

val pp_path_node : Ppx_deriving_runtime.Format.formatter -> path_node -> Ppx_deriving_runtime.unit

val show_path_node : path_node -> Ppx_deriving_runtime.string

type value =

| LVal of lvalue

| Int of int

| Float of float

| String of string

A value in the ANF IR is something immediately available without needing any computation.

type expr =

| Val of value

| Copy of lvalue

| Borrow of mutability * lvalue

| App of value * value list

| UnOp of unary_op * value

| BinOp of binary_op * value * value

| MkData of data_kind * value list

| Fun of func_name

type statement =

| Decl of variable * expr

| Assign of lvalue * value

| EndScope of variable list

type program =

| Jump of span * label * value list

| Stmt of span * statement * program

| Branch of branching

| Block of block_definition * program

| Loop of block_definition

| Empty

| Abort

Return(span, expr) returns the result of expr as the result of the whole program. If expr is a function application, then this represents a tail-call

Jump(span, label, args) jumps to the block with name label with args

Stmt(span, stmt, body) first executes stmt and then executes body. span is the source location of stmt

Branch br is a unified construction for all control flow branching structures, e.g. if/else and pattern matching

Block(def, body) defines a new block with def, and executes body

Loop def defines a recursive block with def, and immediately enter the block

and branching = {

br_src : span;

br_matched : value;

br_branches : (int * program) list;

br_default : program option;

}

branching is a simple switch on ADT label (integer tag). Each ADT label has an associated branch in br_branches, and there is an optional default branch br_default.

branching is used to represent all control flow branching in the source language:

if/else is translated to pattern matching on boolean

pattern matching is decomposed to nested branching

and block_definition = {

blk_label : label;

blk_params : variable list;

blk_body : program;

}

type function_definition = {

func_src : span;

func_name : string;

func_params : variable list;

func_label : label;

func_body : program;

}

func_label represents the point after the function returns. Returning from the function is represented by jumping to func_label.

val gen_var : unit -> int

val gen_label : unit -> int

val reset_generator : unit -> unit

val concat_program : program -> program -> program

val pp_lvalue : Stdlib.Format.formatter -> lvalue -> unit

val pp_value : Stdlib.Format.formatter -> value -> unit

val pp_values : Stdlib.Format.formatter -> value list -> unit

val pp_expr : Stdlib.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val pp_vars : Stdlib.Format.formatter -> int list -> unit

val pp_statement : Stdlib.Format.formatter -> statement -> Ppx_deriving_runtime.unit

val pp_program : Stdlib.Format.formatter -> program -> Ppx_deriving_runtime.unit

val pp_function_definition : Stdlib.Format.formatter -> function_definition -> Ppx_deriving_runtime.unit