Module IR.WasmGen

WASM/WAT code generation from ANF IR. Generates WebAssembly Text Format (.wat).

type wasm_module = {

funcs : wasm_func list;

memories : int;

globals : string list;

}

and wasm_func = {

wf_name : string;

wf_params : (string * wasm_type) list;

wf_result : wasm_type option;

wf_locals : (string * wasm_type) list;

wf_body : wasm_instr list;

}

and wasm_type =

| I32

| I64

| F32

| F64

and wasm_instr =

| WLocalGet of string

| WLocalSet of string

| WLocalTee of string

| WGlobalGet of string

| WGlobalSet of string

| WI32Const of int

| WI64Const of int64

| WF32Const of float

| WF64Const of float

| WCall of string

| WCallIndirect of string

| WBlock of string option * wasm_instr list

| WLoop of string option * wasm_instr list

| WIf of string option * wasm_instr list * wasm_instr list

| WBr of string

| WBrIf of string

| WBrTable of string list * string

| WReturn

| WDrop

| WUnreachable

| WNop

| WI32Add

| WI32Sub

| WI32Mul

| WI32Div_s

| WI32Rem_s

| WI32And

| WI32Or

| WI32Xor

| WI32Shl

| WI32Shr_s

| WI32Eq

| WI32Ne

| WI32Lt_s

| WI32Le_s

| WI32Gt_s

| WI32Ge_s

| WI32Wrap_i64

| WI32Extend_i64_s

| WI64Add

| WI64Sub

| WI64Mul

| WI64Div_s

| WI64Rem_s

| WI64And

| WI64Or

| WI64Xor

| WI64Shl

| WI64Shr_s

| WI64Eq

| WI64Ne

| WI64Lt_s

| WI64Le_s

| WI64Gt_s

| WI64Ge_s

| WF32Add

| WF32Sub

| WF32Mul

| WF32Div

| WF32Eq

| WF32Ne

| WF32Lt

| WF32Le

| WF32Gt

| WF32Ge

| WF64Add

| WF64Sub

| WF64Mul

| WF64Div

| WF64Eq

| WF64Ne

| WF64Lt

| WF64Le

| WF64Gt

| WF64Ge

| I32Load

| I32Store

| I32Load8_s

| I32Store8

| I64Load

| I64Store

| F32Load

| F32Store

| F64Load

| F64Store

| MemorySize

| MemoryGrow

| Select

| Comment of string

val show_wasm_module : wasm_module -> Ppx_deriving_runtime.string

val show_wasm_func : wasm_func -> Ppx_deriving_runtime.string

val show_wasm_type : wasm_type -> Ppx_deriving_runtime.string

val show_wasm_instr : wasm_instr -> Ppx_deriving_runtime.string

WAT Pretty Printer

val pp_wasm_type : Stdlib.Format.formatter -> wasm_type -> unit

val pp_wasm_instr : Stdlib.Format.formatter -> wasm_instr -> unit

val pp_instrs : Stdlib.Format.formatter -> wasm_instr list -> unit

val pp_wasm_func : Stdlib.Format.formatter -> wasm_func -> unit

val pp_wasm_module : Stdlib.Format.formatter -> wasm_module -> unit

Bump Allocator Helpers

val bump_allocator_funcs : wasm_func list

ANF to WASM Compilation

type compile_ctx = {

blocks : (ANF.label, string) Stdlib.Hashtbl.t;

vars : (ANF.variable, string) Stdlib.Hashtbl.t;

funcs : (ANF.func_name, string) Stdlib.Hashtbl.t;

heap_ptr : int Stdlib.ref;

}

Compilation context: maps ANF variables to WASM local names

val mk_compile_ctx : unit -> compile_ctx

val var_wasm_name : compile_ctx -> ANF.variable -> string

val label_wasm_name : compile_ctx -> ANF.label -> string

Memory layout helpers. All values are stored as i32 (pointers or immediate values). Bool: 0 or 1 (stored as i32) Int: i32 Char: i32 (16-bit value) Float: f32/f64 stored as immediate Struct: pointer to heap-allocated struct ADT: pointer to heap-allocated ADT (tag word + payload) Tuple: pointer to heap-allocated tuple

val size_of_type : Semantics.Table.table -> Syntax.ParseTree.typ -> int

How many bytes a type takes in memory (for heap allocation)

val field_offset_of_struct : Semantics.Table.table -> string -> string -> int

Field offset for a named struct field

val comp_value : compile_ctx -> Semantics.Table.table -> ANF.value -> wasm_instr list * wasm_type

Compile expression value: evaluate to a WASM value. Returns (instrs_to_compute_value, wasm_type_of_result)

val comp_path : compile_ctx -> Semantics.Table.table -> ANF.path -> wasm_instr list -> wasm_type -> wasm_instr list * wasm_type

val resolve_field_offset : Semantics.Table.table -> string -> int

val comp_expr : compile_ctx -> Semantics.Table.table -> ANF.expr -> wasm_instr list * wasm_type

Compile an ANF expression to WASM instructions that leave the result on the stack

val comp_values : compile_ctx -> Semantics.Table.table -> ANF.value list -> wasm_instr list * wasm_type list

val comp_binop : ANF.binary_op -> wasm_type -> wasm_instr

val comp_mkdata : compile_ctx -> Semantics.Table.table -> ANF.data_kind -> ANF.value list -> wasm_instr list * wasm_type

val comp_program : compile_ctx -> Semantics.Table.table -> ANF.program -> wasm_instr list

Compile an ANF program segment to WASM instructions

val comp_statement : compile_ctx -> Semantics.Table.table -> ANF.statement -> wasm_instr list

val comp_func : compile_ctx -> Semantics.Table.table -> ANF.function_definition -> wasm_func

Compile a function definition to WASM

FieldByName → Field(index) resolution pass

Before code generation, we resolve all FieldByName occurrences in the ANF to concrete Field indices by scanning the type table. This is a global scan approach: field names are assumed to be unique across all struct types. TODO: carry type information in ANF for precise per-type field resolution.

val build_field_map : Semantics.Table.table -> (string, int) Stdlib.Hashtbl.t

val resolve_field_name : (string, int) Stdlib.Hashtbl.t -> string -> int

val resolve_path_node : (string, int) Stdlib.Hashtbl.t -> ANF.path_node -> ANF.path_node

val resolve_lvalue : (string, int) Stdlib.Hashtbl.t -> ANF.lvalue -> ANF.lvalue

val resolve_value : (string, int) Stdlib.Hashtbl.t -> ANF.value -> ANF.value

val resolve_expr : (string, int) Stdlib.Hashtbl.t -> ANF.expr -> ANF.expr

val resolve_program : (string, int) Stdlib.Hashtbl.t -> ANF.program -> ANF.program

val resolve_statement : (string, int) Stdlib.Hashtbl.t -> ANF.statement -> ANF.statement

val resolve_field_names : Semantics.Table.table -> ANF.function_definition list -> ANF.function_definition list

val comp_module : Semantics.Table.table -> ANF.function_definition list -> wasm_module

Compile an entire program (list of function definitions) to a WASM module

val generate_wat : Semantics.Table.table -> ANF.function_definition list -> string

Generate WAT text from ANF function definitions