Module Semantics.Walker

This module walks AST from Syntax.ParseTree, to do name resolution and type checking in one pass.

Context

type nametbl = (Syntax.ParseTree.variable, Syntax.ParseTree.symbol) Stdlib.Hashtbl.t

type scope = nametbl list

type context = {

table : Table.table;

mutable nametbl : nametbl;

mutable scope : scope;
(*
nametbl::scope
*)

mutable this : Syntax.ParseTree.typ;
(*
type This of one struct or ADT
*)

mutable rety : Syntax.ParseTree.typ;
(*
return type of function
*)

mutable checkloop : int;

}

Helper

Those used in Walker, which is not yet defined in Syntax.ParseTree

Implementation

scope block { stmt_list }

val scope_beg : context -> unit

val scope_end : context -> unit

add variable into nametbl and table

val add_variable : context -> Syntax.ParseTree.symbol -> Syntax.ParseTree.mutability -> Syntax.ParseTree.variable -> Syntax.ParseTree.typ option -> unit

val find_var_opt : context -> Syntax.ParseTree.variable -> Syntax.ParseTree.symbol option

call add_variable

# walk_pattern context pattern typ

typ is the type of expr corresponding to pattern

In let (x, _) = (3, 1.1) , pattern is (x, _), and typ is (i64, f64).

val unsupport_generics : Syntax.ParseTree.typ list -> unit

val walk_type : context -> Syntax.ParseTree.typ -> bool

check TyNamed whether in context.typ

val walk_pattern : context -> Syntax.ParseTree.pattern -> Syntax.ParseTree.typ -> unit

Type Check implemented in walk_expr:

# let ty = walk_expr context expr
  val ty : typ = <...>

ty is the type of expr.

val walk_expr : context -> Syntax.ParseTree.expr -> Syntax.ParseTree.typ

val walk_stmt : context -> Syntax.ParseTree.stmt -> unit

walk stmt As for return type check, please set context.rety

walk_stmt context stmt ;;

When meeting a return-stmt, walk_stmt will compare context.rety with the type of expr

val walk_func_decl : context -> Syntax.ParseTree.func_name -> Syntax.ParseTree.func_arg list -> Syntax.ParseTree.typ -> unit

check func name, and collect type info

# walk_func_decl context name parameters return_type

val walk_func_stmt : context -> Syntax.ParseTree.func_arg list -> Syntax.ParseTree.typ -> Syntax.ParseTree.stmt -> unit

simply walk function's body inductively

val walk_method_intf : context -> Syntax.ParseTree.intf_name option -> Syntax.ParseTree.typ_name -> unit

impl intf for T with methods, check whether intf exists and whether it has such methods

val walk_method_decl : context -> Syntax.ParseTree.typ_name -> Syntax.ParseTree.func_name -> Syntax.ParseTree.func_arg list -> Syntax.ParseTree.typ -> unit

similar to walk_func_decl, but checked in the scope of type T

val walk_method_stmt : context -> Syntax.ParseTree.func_arg list -> Syntax.ParseTree.typ -> Syntax.ParseTree.stmt -> unit

val walk_top : context -> Syntax.ParseTree.top_clause -> unit

Usage

See WalkerTest.ml .