Module ParseTree.NodeId

NodeId contains types for unique ids for various kinds of AST nodes. Can be used to attach extra data to AST nodes, via an external table.

type pattern =

| PatId of int

type expr =

| ExprId of int

type stmt =

| StmtId of int

type impl =

| ImplId of int