Module ParseTree.NodeId

NodeId contains types for unique ids for various kinds of AST nodes. Can be used to attach extra data to AST nodes, via an external table.

type pattern =

| PatId of int

val pp_pattern : Ppx_deriving_runtime.Format.formatter -> pattern -> Ppx_deriving_runtime.unit

val show_pattern : pattern -> Ppx_deriving_runtime.string

type expr =

| ExprId of int

val pp_expr : Ppx_deriving_runtime.Format.formatter -> expr -> Ppx_deriving_runtime.unit

val show_expr : expr -> Ppx_deriving_runtime.string

type stmt =

| StmtId of int

val pp_stmt : Ppx_deriving_runtime.Format.formatter -> stmt -> Ppx_deriving_runtime.unit

val show_stmt : stmt -> Ppx_deriving_runtime.string

type impl =

| ImplId of int

val pp_impl : Ppx_deriving_runtime.Format.formatter -> impl -> Ppx_deriving_runtime.unit

val show_impl : impl -> Ppx_deriving_runtime.string